Working CRUD WinForm Application
Path to executable is in the debug folder. It will eventually have an installer for you to put in your server IP, username, and password for now if you try it will not run unless on my computer so that it gets the right IP address. This application is able to create, edit and delete packages with it automatically converting weights to pounds in the MySql database and creating the created date and sent date for when the package is made or status is updated to sent.